GVT-2916 & others: Move number formatter back to its original place
