--- infra/src/main/kotlin/fi/fta/geoviite/api/formatting/NumberFormatting.kt
@@ -1,13 +1,0 @@
-package fi.fta.geoviite.api.formatting
-
-import java.math.BigDecimal
-import java.math.RoundingMode
-
-fun splitBigDecimal(number: BigDecimal, decimalPlaces: Int = 3): Pair<Int, Int> {
-    val wholePart = number.toBigInteger().toInt()
-    val fractionalPart = number.subtract(BigDecimal(wholePart))
-    val scaledFractionalPart =
-        fractionalPart.setScale(decimalPlaces, RoundingMode.DOWN).movePointRight(decimalPlaces).toInt()
-
-    return Pair(wholePart, scaledFractionalPart)
-}
